move tokenizer and the rest of shared
