+++ lib/discourse_ai/chat_message_classificator.rb
