+++ lib/discourse_ai/classificator.rb
