+++ lib/discourse_ai/database/connection.rb
