+++ lib/discourse_ai/post_classificator.rb
