+++ lib/discourse_ai/tokenizer/multilingual_e5_large_tokenizer.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module DiscourseAi
+  module Tokenizer
+    class MultilingualE5LargeTokenizer < BasicTokenizer
+      def self.tokenizer
+        @@tokenizer ||=
+          Tokenizers.from_file("./plugins/discourse-ai/tokenizers/multilingual-e5-large.json")
+      end
+    end
+  end
+end
